improve conditions in code; fix typo
