+++ indy_node/test/upgrade/helper.py
@@ -1,28 +1,39 @@
-from plenum.common.util import randomString
-from stp_core.loop.eventually import eventually
-from plenum.test.helper import waitForSufficientRepliesForRequests
-from plenum.test import waits as plenumWaits
-from plenum.common.types import f
-from plenum.common.constants import TXN_TYPE, DATA, VERSION
-from indy_common.constants import NODE_UPGRADE, ACTION
-from indy_client.client.wallet.upgrade import Upgrade
-from indy_node.server.upgrader import Upgrader
-from indy_node.server.upgrade_log import UpgradeLog
-from indy_node.utils.node_control_tool import NodeControlTool
-from indy_common.config_util import getConfig
+import functools
+import json
+import multiprocessing
+import os
+import socket
+import subprocess
 from datetime import datetime
 from typing import List, Tuple
+
 import dateutil.tz
-import subprocess
-import os
-import multiprocessing
-import socket
-import json
-import functools
+from plenum.common.constants import TXN_TYPE, DATA, VERSION
+from plenum.common.types import f
+from plenum.common.util import randomString
+from plenum.test import waits as plenumWaits
+from plenum.test.helper import waitForSufficientRepliesForRequests
 from stp_core.common.log import getlogger
+from stp_core.loop.eventually import eventually
+
+from indy_client.client.wallet.upgrade import Upgrade
+from indy_common.config_util import getConfig
+from indy_common.constants import NODE_UPGRADE, ACTION
+from indy_node.server.upgrade_log import UpgradeLog
+from indy_node.server.upgrader import Upgrader
+from indy_node.test.helper import TestNode
+from indy_node.utils.node_control_tool import NodeControlTool
 
 config = getConfig()
 logger = getlogger()
+
+
+class TestNodeNoProtocolVersion(TestNode):
+    def processNodeRequest(self, request, frm):
+        if request.protocolVersion is not None:
+            raise ValueError('Do not understand what protocolVersion is!!!')
+        super().processNodeRequest(request, frm)
+
 
 def sendUpgrade(client, wallet, upgradeData):
     upgrade = Upgrade(**upgradeData, trustee=wallet.defaultId)
@@ -96,20 +107,19 @@ class NodeControlToolExecutor:
         transform(self.tool)
         self.p = multiprocessing.Process(target=self.tool.start)
         self.p.start()
-        logger.debug("NCTProcess was started with pid: {}". format(self.p.pid))
+        logger.debug("NCTProcess was started with pid: {}".format(self.p.pid))
 
     def stop(self):
-        logger.debug("Send stop to NCTProcess with pid: {}". format(self.p.pid))
+        logger.debug("Send stop to NCTProcess with pid: {}".format(self.p.pid))
         self.tool.server.close()
         self.p.terminate()
-        # check that process with NetControlTool.start function really stop.
+        # check that process with NodeControlTool.start function really stop.
         # process.terminate() just send SIGTERM and is not guarantee that process stops
         while self.p.is_alive():
-            logger.debug("NCTProcess still alive, with pid: {}". format(self.p.pid))
+            logger.debug("NCTProcess still alive, with pid: {}".format(self.p.pid))
             # while process is still alive, join with main process and wait
             self.p.join(3)
-        logger.debug("NCTProcess must be stopped, with pid: {}". format(self.p.pid))
-
+        logger.debug("NCTProcess must be stopped, with pid: {}".format(self.p.pid))
 
 
 def composeUpgradeMessage(version):
@@ -195,6 +205,7 @@ def check_node_do_not_sent_acknowledges_upgrade(
     check_ledger_after_upgrade(node_set, allowed_actions,
                                ledger_size, expected_version,
                                node_ids=node_ids)
+
 
 def check_ledger_after_upgrade(
         node_set,
