+++ indy_node/test/upgrade/test_node_upgrade_no_protocol_version.py
@@ -2,20 +2,12 @@ import pytest
 
 import indy_node
 from indy_common.constants import COMPLETE
-from indy_node.test.helper import TestNode
 from indy_node.test.upgrade.helper import populate_log_with_upgrade_events, \
-    check_node_sent_acknowledges_upgrade
+    check_node_sent_acknowledges_upgrade, TestNodeNoProtocolVersion
 
 whitelist = ['unable to send message']
 
 version = indy_node.__metadata__.__version__
-
-
-class TestNodeNoProtocolVersion(TestNode):
-    def processNodeRequest(self, request, frm):
-        if request.protocolVersion is not None:
-            raise ValueError('Do not understand what protocolVersion is!!!')
-        super().processNodeRequest(request, frm)
 
 
 @pytest.fixture(scope="module")
